requirements.txt
wordcloud~=1.8.1
translators~=4.9.5
numpy~=1.21.2
Pillow~=8.3.2